refactor contract tests data model into dedicated library
