+++ contract-tests/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(data-model)
 add_subdirectory(sse-contract-tests)
 add_subdirectory(client-contract-tests)
 add_subdirectory(server-contract-tests)
