+++ contract-tests/data-model/src/data_model.cpp
@@ -1,4 +1,4 @@
-#include "definitions.hpp"
+#include "data_model/data_model.hpp"
 
 EvaluateFlagParams::EvaluateFlagParams()
     : valueType{ValueType::Unspecified}, detail{false} {}
