+++ contract-tests/server-contract-tests/include/entity_manager.hpp
@@ -5,8 +5,7 @@
 #include <launchdarkly/logging/logger.hpp>
 #include <launchdarkly/server_side/client.hpp>
 
-#include "client_entity.hpp"
-#include "definitions.hpp"
+#include <data_model/data_model.hpp>
 
 #include <memory>
 #include <mutex>
@@ -17,7 +16,7 @@
 class EventOutbox;
 
 class EntityManager {
-    std::unordered_map<std::string, ClientEntity> entities_;
+    //std::unordered_map<std::string, ClientEntity> entities_;
 
     std::size_t counter_;
     boost::asio::any_io_executor executor_;
