--- contract-tests/server-contract-tests/src/definitions.cpp
@@ -1,6 +1,0 @@
-#include "definitions.hpp"
-
-EvaluateFlagParams::EvaluateFlagParams()
-    : valueType{ValueType::Unspecified}, detail{false} {}
-
-CommandParams::CommandParams() : command{Command::Unknown} {}
